+++ qatrack/qatrack_core/management/commands/installfixtures.py
@@ -10,8 +10,8 @@ class Command(BaseCommand):
     help = 'Loads all default fixtures from fixtures/defaults/'
 
     def handle(self, *args, **kwargs):
-        # Construct the path to the fixtures directory
-        fixtures_pattern = os.path.join(settings.BASE_DIR, 'fixtures', 'defaults', '*', '*.json')
+        base_dir = getattr(settings, 'BASE_DIR', os.path.abspath(os.path.join(settings.PROJECT_ROOT, '..')))
+        fixtures_pattern = os.path.join(base_dir, 'fixtures', 'defaults', '*', '*.json')
         fixtures = glob.glob(fixtures_pattern)
         
         if not fixtures: